Implement PII message identification (UXPROD-336)
We use JSONPath to identify personally identifiable information (PII) within the bodies of messages that are transmitted via AES. When encountered, we set the "pii" metadata field to `true`, otherwise it is `false`. The list of JSONPaths is not comprehensive. It will need to be expanded as we better understand what PII is and how to identify it within FOLIO.