MODAES-4: enable decompression for compressed client side data
Okapi can route compressed data destined for the client, but mod-aes
does not decompress the data, so the compressed data is sent on to kafka
or logged and is not usable. We now allow the server to decompress
client side data, if received compressed with a "Content-Encoding"
header with the compression method, likely gzip. Also enabled
compression support in the server, probably not necessary given the
current APIs.