MODAES-7: URL encode the CQL query
nginx does not like spaces in the query, probably anywhere in the URL as
well, so we need to encode the CQL passed in the `query` query
parameter. It is difficult to write a test for this, but all CQL should
be encoded anyway, to be on the safe side. The rules test was updated to
include the full path we use for configuration retrieval. When set, the
test fails without encoding. Vert.x appears to close the connection
before it gets to the response code. The test works fine with the CQL
encoded and the full path the configuration specified.