MODAES-5: Fix a NPE
When there is no requirement to send the content type header, such as on
a GET, we need test to see if the header is present before operating on
it. Added a test that does not pass a content type header which fails
with the old code (NPE) and succeeds with this commit.