Timing problems with mockito and the compression test
Simplify the test by making the stub lenient and not doing an explicit
undeploy. The NPE is still prevented by the stub, which will eventually
be called by undeploy, but can occur after Mockito is notified that the
test has completed. Without lenient set, it will report a stub error.
Waiting after the undeploy is possible, but seemed arbitrary and not
really necessary for the test.