MODAES-13: fix WebClient memory leak
WebClients were being created whenever the configuration API was called,
but these clients were not being closed. Since they were not closed, the
underlying connection objects were not cleaned up. Eventually, the
module will run out of memory.

We now create one WebClient when the verticle is started. This client is
reused for all configuration API calls. The client is closed when the
verticle is stopped.