MODAES-15 and MODAES-16: reduce garbage and simplify flow
MODAES-15: Removed creation of objects that were either not used or were
used temporarily, but could have been reused. Cleaned up some async
workflows so that they are more concise. Refactored some code that can
now take advantage of Java 11. Added config so that we can enable async
logging via Log4j2 without having to rebuild the fat jar. Once deployed,
async logging can be enabled by supplying
`-Dlog4j2.contextSelector=org.apache.logging.log4j.core.async.AsyncLoggerContextSelector`
at startup.

MODAES-16: instead of using a timeout check, which can lead to a state where there is nothing in the cache and many asynchronous requests fire configuration requests, we now use `computeIfAbsent` to retrieve the rules from the configuration interface and store the `Future` in the map. If the `Future` succeeds, we set a Vert.x timer that will remove the rules `Future` when the timer completes. If the `Future` fails, we remove the rules `Future` from the map so the next request will trigger a request. Since this is all done in a `ConcurrentHashMap`, the configuration request is only performed once per cache period, assuming success and constant traffic. When the `Future` is removed from the map all requests will get the `Future` that is stored in the map via the first request that finds the key to be absent in the map. When the `Future` completes successfully, all waiting requests will complete with the retrieved rules. Future requests will get the completed `Future` from the map. In this way, the AES service does not require code changes to use this fix since we are still returning a `Future<Collection<RoutingRule>>`.